Propagate timestamps from CallInfo to TestReport objects
TestReport objects stored only the information about stage duration.
This is not enough to correlate pytest stages with external events.
Timestamps are already there in CallInfo. They just needed to be
propagated to TestReport objects, so that they can be read when
TestReports are exported externally (for example with --reportlog
option)