doc: remove unwanted implication
Returning None does not mean that all future calls to the underlying
platform will result in a return value of None. It is a certainty for win32
and Emscripten platforms but only a possibility for other platforms.
The documentation should only state that the *current* call to the
underlying platform was unsuccessful.