refactor: use division operator to join paths
Starting with `resolve_package_path` and its associated tests,
this refactoring seeks to make path concatenation more
readable and consistent.

To be consistent, all uses of `.joinpath()` need to be replaced
with the division operator or the other way around. I have chosen
to standardise the use of the division operator because
   1) it is already used in the tests
   2) it is more succinct and readable
   3) casual profiling with `%timeit` didn't reveal a performance loss
           division operator: 273 µs ± 86.6 µs
                    `.joinpath()`: 269 µs ± 99.4 µs

Eventual consistency within and across all files is the ultimate goal.
This is not unachievable. There are currently 428 uses of `.joinpath()`
across 45 files. Counting assignment expressions only, there are 111
uses of the division operator across 23 files. This is fewer. However,
even if there is a usage pattern of 4:1 against the division operator,
I do not think that makes standardising it an unwise choice.