Fix PytestPluginManager._is_in_confcutdir on Windows
On Windows two paths might be in different drives, so if a path is not one of the parents of the confcutdir, it does not necessarily mean it is inside the confcutdir either.

Use `Path.is_relative_to` instead which directly answers that question for us.

The previous implementation could cause pytest to inspect all the parent paths of a path from one drive if pytest was invoked from another drive.

Moved the implementation to a function because it would be too convoluted to test the `PytestPluginManager._is_in_confcutdir` method directly.