fixtures: exclude direct-param fixtures from --fixtures-per-test output
Fix #11295 by excluding from the --fixtures-per-test output any 'pseudo
fixture' that results from directy parametrizing a test e.g. with
``@pytest.mark.parametrize``.

The justification for removing these fixtures from the report is that

a) They are unintuitive. Their appearance in the fixtures-per-test
report confuses new users because the fixtures created via
``@pytest.mark.parametrize`` do not confrom to the expectations
established in the documentation; namely, that fixtures are
- richly reusable
- provide setup/teardown features
- created via the ``@pytest.fixture` decorator

b) They are an internal implementation detail. It is not the explicit
goal of the direct parametrization mark to create a fixture; instead,
pytest's internals leverages the fixture system to achieve the explicit
goal: a succinct batch execution syntax. Consequently, exposing the
fixtures that implement the batch execution behaviour reveal more about
pytest's internals than they do about the user's own design choices and
test dependencies.