Bump PyPy runtime to v3.9 @ GHA
PyPy 3.8 has a flaky bug in the garbage collector that is not going to
be fixed there since that version is EoL, but newer versions have it
addressed.

The problem manifests itself in CI as follows:

    ```console
    TypeError: expected some sort of stmt, but got <_ast.Load object ...>
    ```

This patch makes use of the PyPy version that is still supported
and contains a fix for the said bug.

Ref: https://www.pypy.org/posts/2024/03/fixing-bug-incremental-gc.html

Resolves #11771.