Remove quotes from example tmp_path_retention_policy
Previously the docs contained an example for `tmp_path_retention_policy`
which included quotes around the value. If the configuration was
actually set with this value. The following error would be encountered

`INTERNALERROR> ValueError: tmp_path_retention_policy must be either all, failed, none. Current input: "all".`

This is because the value of the field was not being unwrapped from
the quotes when parsed.

This commit removes the quotes from the example so that if used in the
configuration file, no error will occur.