Amend docs to better report new behavior:
When I started this feature, I expected passing the "tee" flag
to "syscap" would copy output directly to the calling terminal, to the
original ORIGINAL sys.stdout/sys.stderr. That's not how it works-
it passes output to whatever capture was set earlier! This actually
makes the feature more flexible because it enables tee-like behavior as
well as pytests reporting behavior. Awesome!