Remove 'pip' command hint from release announcement
While of course `pip` is still the most used tool to install packages, currently there are a lot of alternatives (`poetry`, `uv`, etc), so it feels a bit dated to suggest `pip install --upgrade` directly on the release announcement.

Also, all the other release notes do not mention `pip` too, so it makes sense to remove it from the patch release announcement.