Fix compatibility with Twisted 25
The issue arises because `Failure.__init__` in Twisted 25 can receive a non-None `exc_value` while `exc_tb` is `None`. In such cases, `ExceptionInfo[BaseException].from_exc_info` fails, as it expects a traceback when `sys.exc_info()` returns a tuple. This leads to the error message `'NoneType' object is not iterable`.

Adjusted the `Failure.__init__` wrapper's logic to ensure we always have a valid traceback.

Tested with `twisted 24.11.0` and `25.5.0`.

Fixes #13497