build: replace outdated *-latest runners with specific ones
This change unlocks the current pipelines because it will allow
use of newer version of 3.9 which is required by tox.

In general is better not to use `-latest` runners because they are
very often outdated versions, causing problems. As github does not
update the name of the runners very often, it is easier to use
specific ones. This also prevents surprises when github is changing
the runner versions as they do this gradually and -latest ones might
point to different runners in different jobs.

Related: https://github.com/tox-dev/tox/issues/3565
Reference: https://github.com/actions/runner-images