runner: correctly cleanup item _request/funcargs if an exception was reraised during call (e.g. KeyboardInterrupt)
In my test suite, I have some objects that rely on garbage collection to
be cleaned up correctly. This is not amazing, but it's how the code is
structured for now. If I interrupt tests with Ctrl+C, or by manually
raising KeyboardInterrupt in a test, these objects are not cleaned up
any more.

call_and_report re-raises Exit and KeyboardInterrupt, which breaks the
cleanup logic in runtestprotocol that unsets the item funcargs (which is
where my objects end up living as references, as they're passed in as
fixtures).

By just wrapping the entire block with try: ... finally: ..., cleanup
works again as expected.

Co-authored-by: Ran Benita <ran@unusedvar.com>