Flush stdout and stderr in pytester.run (#13695)
Attempted to make `test_faulthandler.py::test_timeout` more reliable on CI, but ultimately this was not accomplished.

However, it seems flushing `stdout` and `stderr` during `pytester.run` seems the right thing to do, so it was decided to leave the calls and only skip the test in more specific scenarios that were failing frequently on CI (see #13695 for discussion history).

We should extend the skip list as needed.

Follow-up on #13684.
Partially addresses #7022.