main: put parametrization in a separate field on CollectionArgument
This is for the benefit of the next commit. That commit wants to check
whether a CollectionArgument is subsumed by another. According to pytest
semantics:
    `test_it.py::TestIt::test_it[a]` subsumed by `test_it.py::TestIt::test_it`
However the `parts` are
    ["TestIt", test_it[a]"]                       ["TestIt", test_it"]
which means a simple list prefix cannot be used. By splitting the
parametrization `"[a]"` part to its own attribute, it can be handled
cleanly.

I also think this is a reasonable change regardless. We'd probably want
something like this when the "collection structure contains
parametrization" TODO is tackled.