fixtures: fix closure computation with indirect override chains
Continuation of 72ae3dbf9608a70df683d081cf53146a8d79f1ea.

The previous fix was a minimal change to make the existing code at least
consider fixtures overrides, and handles the common case (direct
override chains), but still wasn't correct, as it didn't handle
override chains involving an intermediary (not the overridden) fixture.

To make this work, the algorithm needs to change. Now instead of a
simple breadth-first search, we do a depth-first search, which more
closely simulates the runtime behavior, and allows us the to check the
"stack" of fixtures so we can use the correct fixture index (depth in
the override chain).

This is more expensive but should be OK.

Refs #13773.